[GOHW-4]: feat: add sqlc code and repository

+++ loms/migrations/20240924020802_create_items_table.sql
@@ -1,8 +1,8 @@
 -- +goose Up
 -- +goose StatementBegin
 CREATE TABLE items (
-    id SERIAL PRIMARY KEY,
-    order_id INTEGER REFERENCES orders(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY,
+    order_id BIGINT REFERENCES orders(id) ON DELETE CASCADE,
     sku INTEGER NOT NULL,
     count SMALLINT NOT NULL
 );
